Fixing the test suite
Couple problems were unveiled:
- the records handling of isUpdating requires more heavier refactoring,
where the live record arrays are based on a common structure and there
is just a thin wrapper around this structure. This is only needed
because fetching/updating an array, will affect all associated arrays
- polymorphic relationships creates a problem, becuase they can specify
type X, but the actual record is of type Y, where Y is subtype of X. If
we are to allow multi-records, we need to be able to specify for
relationships a mapping for Y -> Y', Z -> Z'.